# swagger.yaml
openapi: 3.0.0
info:
  title: Mi E-Commerce
  version: 2.0.0
  description: |-
    Esta API esta orientada a interactuar con el front de una pagina web de E-Commerce del grupo Cencosud SA.
    Git Repo: https://github.com/rodrigobenta/segundoSprint  
servers:
  - url: http://localhost:3000/api/v2
    description: Entorno Desarrollo
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /products/{id}/pictures:
    get:
      tags:
        - pictures
      summary: Lista las fotos del producto identificado con id
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: number
          required: true
          description: Id del producto
          
      responses: 
        '200':
          description: Fotos listadas con exito
          content:
            application/json:
              schema:
                properties:
                  list:
                    type: array
                    items:
                      type: object
                    properties:
                      id:
                        type: number
                        
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                properties:
                  msg:
                    type: string
                    
        '500':
          description: Server error
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
  /pictures:
    post:
      tags: 
        - pictures
      summary: Crea una nueva foto para un producto 
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                url:
                  type: string
                description:
                  type: string
                fk_id_product:
                  type: number
      
      responses: 
        '201':
          description: Foto creada con exito
          content:
            application/json:
              schema:
                properties:
                      id:
                        type: number
                      url:
                        type: string
                      description:
                        type: string
                      fk_id_product:
                        type: string
  
        '400':
          description: bad request
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
        
        '500':
          description: Server error
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
    
    get:
      tags: 
        - pictures
      summary: Busca las fotos de un producto identificado con un Id 
      security:
        - bearerAuth: []
      parameters:
        - name: product
          in: query
          schema:
            type: string
          required: false
          description: Id del producto
      
      responses: 
        '200':
          description: Fotos listadas con exito
          content:
            application/json:
              schema:
                properties:
                  
                      id:
                        type: number
                      url:
                        type: string
                      description:
                        type: string

        '400':
          description:  bad request
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
        
        '500':
          description: Server error
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
  /pictures/{id}:                 
    delete:
      tags: 
        - pictures
      summary: Borra la foto identificada con id 
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Id de la foto
      
      responses: 
        '200':
          description: Foto eliminada con exito
          content:
            application/json:
              schema:
                properties:
                      id:
                        type: number
                      url:
                        type: string
                      description:
                        type: string
                      
        '400':
          description:  bad request
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
        
        '500':
          description: Server error
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string

    put:
      tags: 
        - pictures
      summary: Edita una foto identificada con Id
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Id de la foto
      requestBody:
        required: true
        content:
          application/json:
            schema:
                properties:
                    url:
                      type: string
                    description:
                      type: string
                    fk_id_product:
                      type: number

      responses: 
        '200':
          description: Foto editada con exito
          content:
            application/json:
              schema:
                properties:
                  picture_id:
                    type: number
                  url:
                    type: string
                  description:
                    type: string
                      
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                properties:
                  msg:
                    type: string
                  ok:
                    type: boolean
                    
        '404':
          description: Foto no encontrada
          content:
            application/json:
              schema:
                properties:
                  msg:
                    type: string
                  ok:
                    type: boolean
        '500':
          description: Server error
          content:
            application/json:
              schema:
                properties:
                  msg:
                    type: string
                  ok:
                    type: boolean
                    
    get:
      tags: 
        - pictures
      summary: Busca la foto identificada con Id
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: busca producto
      
      responses: 
        '200':
          description: Foto encontrada con exito
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: number
                  url:
                    type: string
                  description:
                    type: string
                  fk_id_product:
                    type: number
                      
        '400':
          description:  bad request
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
        '404':
          description:  Not found
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
        
        '500':
          description:  server error
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
  /carts/{id}:
    get:
      tags: 
        - carts
      summary: Busca el carrito del usuario identificado con id
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Id del usuario
              
      responses: 
        '200':
          description: carrito encontrado con exito
          content:
            application/json:
              schema:
                properties:
                  carrito:
                      type: array
                      items:
                        type: object 
                        properties:
                          product:
                            type: object
                            properties:
                              id:
                                type: number
                              title:
                                type: string
                              price:
                                type: number
                              description:
                                type: string
                              image:
                                type: string
                              gallery:
                                type: array
                                items:
                                  type: object
                                  properties:
                                    picture_id:
                                      type: number
                                    picture_url:
                                      type: string
                                    description:
                                      type: string
                              category:
                                type: string
                              mostwanted:
                                type: string
                              stock:
                                type: number
                          quantity:
                            type: number

        '404':
          description:  Usuario no encontrado
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string

        '500':
          description: Server error
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
                    
    put:
      tags: 
        - carts
      summary: Actualiza el carrito del usuario identificado con id
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Id del usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              items:
                type: object
                properties:
                  product:
                    type: number
                  quantity:
                    type: number
              
      responses: 
        '200':
          description: Actualizado con exito
          content:
            application/json:
              schema:
                properties:
                  user:
                    type: number
                  gallery:
                        type: array
                        items:
                          type: object 
                          properties:
                            product:
                              type: number
                            quantity:
                              type: number
        '500':
          description: Server error
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
                    
        '404':
          description:  Usuario no encontrado
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string